+++ src/Deveel.Math.Core/Deveel.Math.Core.csproj
@@ -33,15 +33,13 @@
     <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimal.cs" Link="Math\BigDecimal.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimalMath.cs" Link="Math\BigDecimalMath.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimal_Convertible.cs" Link="Math\BigDecimal_Convertible.cs" />
-    <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimal_Functions.cs" Link="Math\BigDecimal_Functions.cs" />
-    <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimal_Operators.cs" Link="Math\BigDecimal_Operators.cs" />
-    <Compile Include="..\Deveel.Math\Deveel.Math\BigDecimal_Parse.cs" Link="Math\BigDecimal_Parse.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigInteger.cs" Link="Math\BigInteger.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigIntegerMath.cs" Link="Math\BigIntegerMath.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigMath.cs" Link="Math\BigMath.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BitLevel.cs" Link="Math\BitLevel.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\CharHelper.cs" Link="Math\CharHelper.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\Conversion.cs" Link="Math\Conversion.cs" />
+    <Compile Include="..\Deveel.Math\Deveel.Math\DecimalString.cs" Link="Math\DecimalString.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\Division.cs" Link="Math\Division.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\Elementary.cs" Link="Math\Elementary.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\Logical.cs" Link="Math\Logical.cs" />
