+++ src/Deveel.Math/Deveel.Math.csproj
@@ -158,10 +158,8 @@
   <ItemGroup>
     <Compile Include="Deveel.Math\BigDecimal.cs" />
     <Compile Include="Deveel.Math\BigDecimalMath.cs" />
+    <Compile Include="Deveel.Math\DecimalString.cs" />
     <Compile Include="Deveel.Math\BigDecimal_Convertible.cs" />
-    <Compile Include="Deveel.Math\BigDecimal_Functions.cs" />
-    <Compile Include="Deveel.Math\BigDecimal_Operators.cs" />
-    <Compile Include="Deveel.Math\BigDecimal_Parse.cs" />
     <Compile Include="Deveel.Math\BigInteger.cs" />
     <Compile Include="Deveel.Math\BigIntegerMath.cs" />
     <Compile Include="Deveel.Math\BigMath.cs" />
